Transport: DRY up all the tryConnect implementations
