+++ src/common/lib/transport/websockettransport.ts
@@ -1,6 +1,6 @@
 import Platform from 'common/platform';
 import * as Utils from '../util/utils';
-import Transport, { TryConnectCallback } from './transport';
+import Transport from './transport';
 import Defaults from '../util/defaults';
 import Logger from '../util/logger';
 import ProtocolMessage from '../types/protocolmessage';
@@ -30,40 +30,6 @@ class WebSocketTransport extends Transport {
 
   static isAvailable() {
     return !!Platform.Config.WebSocket;
-  }
-
-  static tryConnect(
-    connectionManager: ConnectionManager,
-    auth: Auth,
-    params: TransportParams,
-    callback: TryConnectCallback
-  ) {
-    const transport = new WebSocketTransport(connectionManager, auth, params);
-
-    let transportAttemptTimer: NodeJS.Timeout | number;
-    const errorCb = function (this: { event: string }, err: ErrorInfo) {
-      clearTimeout(transportAttemptTimer);
-      callback({ event: this.event, error: err });
-    };
-
-    const realtimeRequestTimeout = connectionManager.options.timeouts.realtimeRequestTimeout;
-    transportAttemptTimer = setTimeout(() => {
-      transport.off(['preconnect', 'disconnected', 'failed']);
-      transport.dispose();
-      errorCb.call(
-        { event: 'disconnected' },
-        new ErrorInfo('Timeout waiting for transport to indicate itself viable', 50000, 500)
-      );
-    }, realtimeRequestTimeout);
-
-    transport.on(['failed', 'disconnected'], errorCb);
-    transport.on('preconnect', function () {
-      Logger.logAction(Logger.LOG_MINOR, 'WebSocketTransport.tryConnect()', 'viable transport ' + transport);
-      clearTimeout(transportAttemptTimer);
-      transport.off(['failed', 'disconnected'], errorCb);
-      callback(null, transport);
-    });
-    transport.connect();
   }
 
   createWebSocket(uri: string, connectParams: Record<string, string>) {
