+++ src/platform/web/lib/transport/xhrpollingtransport.js
@@ -1,6 +1,4 @@
 import * as Utils from '../../../../common/lib/util/utils';
-import ErrorInfo from '../../../../common/lib/types/errorinfo';
-import Logger from '../../../../common/lib/util/logger';
 import Platform from '../../../../common/platform';
 import CometTransport from '../../../../common/lib/transport/comettransport';
 import XHRRequest from './xhrrequest';
@@ -17,35 +15,6 @@ var XHRPollingTransport = function (connectionManager) {
 
   XHRPollingTransport.isAvailable = function () {
     return Platform.Config.xhrSupported && Platform.Config.allowComet;
-  };
-
-  XHRPollingTransport.tryConnect = function (connectionManager, auth, params, callback) {
-    var transport = new XHRPollingTransport(connectionManager, auth, params);
-
-    var transportAttemptTimer;
-    var errorCb = function (err) {
-      clearTimeout(transportAttemptTimer);
-      callback({ event: this.event, error: err });
-    };
-
-    var realtimeRequestTimeout = connectionManager.options.timeouts.realtimeRequestTimeout;
-    transportAttemptTimer = setTimeout(function () {
-      transport.off(['preconnect', 'disconnected', 'failed']);
-      transport.dispose();
-      errorCb.call(
-        { event: 'disconnected' },
-        new ErrorInfo('Timeout waiting for transport to indicate itself viable', 50000, 500)
-      );
-    }, realtimeRequestTimeout);
-
-    transport.on(['failed', 'disconnected'], errorCb);
-    transport.on('preconnect', function () {
-      Logger.logAction(Logger.LOG_MINOR, 'XHRPollingTransport.tryConnect()', 'viable transport ' + transport);
-      clearTimeout(transportAttemptTimer);
-      transport.off(['failed', 'disconnected'], errorCb);
-      callback(null, transport);
-    });
-    transport.connect();
   };
 
   XHRPollingTransport.prototype.toString = function () {
