+++ src/platform/web/lib/transport/xhrstreamingtransport.js
@@ -1,7 +1,5 @@
 import * as Utils from '../../../../common/lib/util/utils';
 import CometTransport from '../../../../common/lib/transport/comettransport';
-import ErrorInfo from '../../../../common/lib/types/errorinfo';
-import Logger from '../../../../common/lib/util/logger';
 import Platform from '../../../../common/platform';
 import XHRRequest from './xhrrequest';
 
@@ -17,37 +15,6 @@ var XHRStreamingTransport = function (connectionManager) {
 
   XHRStreamingTransport.isAvailable = function () {
     return Platform.Config.xhrSupported && Platform.Config.streamingSupported && Platform.Config.allowComet;
-  };
-
-  XHRStreamingTransport.tryConnect = function (connectionManager, auth, params, callback) {
-    var transport = new XHRStreamingTransport(connectionManager, auth, params);
-
-    var transportAttemptTimer;
-    var errorCb = function (err) {
-      clearTimeout(transportAttemptTimer);
-      callback({ event: this.event, error: err });
-    };
-
-    var realtimeRequestTimeout = connectionManager.options.timeouts.realtimeRequestTimeout;
-    transportAttemptTimer = setTimeout(function () {
-      transport.off(['preconnect', 'disconnected', 'failed']);
-      transport.dispose();
-      errorCb.call(
-        { event: 'disconnected' },
-        new ErrorInfo('Timeout waiting for transport to indicate itself viable', 50000, 500)
-      );
-    }, realtimeRequestTimeout);
-
-    transport.on(['failed', 'disconnected'], errorCb);
-
-    transport.on('preconnect', function () {
-      Logger.logAction(Logger.LOG_MINOR, 'XHRStreamingTransport.tryConnect()', 'viable transport ' + transport);
-      clearTimeout(transportAttemptTimer);
-      transport.off(['failed', 'disconnected'], errorCb);
-      callback(null, transport);
-    });
-
-    transport.connect();
   };
 
   XHRStreamingTransport.prototype.toString = function () {
